incomplete version 1
This version was built to perform simple calculations (a + b = c) however when I tried to add continuous calculations as a feature it became clear that the original code was not fit for purpose.  Version 2  will likely involve a complete reconstruction.

Also - The font I chose for the display fit the theme of an oldschool 7-segment LED display, but it displayed a decimal-point and a minus/subtraction symbol as the same thing. More research needed!